task/requirements.txt
apache-airflow==2.7.0
kafka-python==2.0.2
pandas==2.0.3
requests==2.31.0